modified useContext problem, add a Counter for shoppingcart
