+++ components/Navigation/index.js
@@ -1,8 +1,12 @@
 import React from "react";
 import styled from "styled-components";
 import Link from "next/link";
+import { useContext } from "react";
+import { CartContext } from "../../pages/_app";
 
 export default function Navigation() {
+
+  const context=useContext(CartContext)
   return (
     <>
       <StyledListItem>
@@ -15,6 +19,7 @@ export default function Navigation() {
       </StyledListItem>
       <StyledListItem>
         <Link href="http://localhost:3000/shopping-cart">shoppingcart</Link>
+        <div>{context.totalAmount}</div>
       </StyledListItem>
       <StyledListItem>
         <Link href="#">profil</Link>
